agents_catalog/25-scientific-poster-extraction-agent/requirements.txt
urllib3
strands-agents 
strands-agents-tools
uv